fix(web): collection + library writes go through plain-primitive maps
The web build couldn't write collections or save to library. Two GitLive/JS
serializer faults, neither of which affects desktop/Android:

  1. A Kotlin Long handed to the Firebase JS SDK is rejected
     ("Unsupported field value: a custom Long object") — hit by
     toLibraryProjection's total_tracks / projection_fetched_at.
  2. set(CollectionDocument) runs the generated serializer, whose List
     descriptor reports elementsCount==1, so an EMPTY `albums` is written as
     [null]. That document then fails every future read, which takes down the
     whole collections list. Confirmed in prod: users/<uid>/collections/
     "Long Fix Test" had albums=[null].

Fix: every shared-code Firestore write is now a hand-built Map<String, Any?> of
String/Int/Double/Boolean/null (+ List/Map of those). Numbers are Int or Double,
never Long; the `albums` array is List<Map>, which goes through GitLive's
size-aware list serializer so [] stays []. Reads are unchanged — models keep
Long and GitLive coerces Number -> Long on every platform.

  - db/FirestoreMap.kt: isFirestoreRawValueSafe / isFirestoreRawMapSafe guard.
  - AlbumMapper.collectionEntryToFirestoreMap; toLibraryProjection emits Int/Double.
  - AlbumCollectionRepository.createCollection / updateCollectionByName and
    CollectionAlbumRepository add/remove/reorder/clear -> primitive maps.
  - FirestoreMapTest + AlbumMapperTest updated. desktopTest 88/0; 3 targets compile.

Album/artist/individual-track document writes still use set(model) — separate
commit. Repair the existing [null] doc with:
  python3 scripts/scan_firestore_health.py --cred serviceAccountKey.json --repair-null-albums

+++ composeApp/src/desktopTest/kotlin/io/github/peningtonj/recordcollection/db/mapper/AlbumMapperTest.kt
@@ -51,7 +51,7 @@ class AlbumMapperTest {
             artistsJson = p["artists"] as String,
             releaseDate = p["release_date"] as String,
             albumType = p["album_type"] as String,
-            totalTracks = p["total_tracks"] as Long,
+            totalTracks = (p["total_tracks"] as Int).toLong(),
             spotifyId = p["spotify_id"] as String,
             spotifyUri = p["spotify_uri"] as String,
             imageUrl = p["image_url"] as String?,
